+++ dart/lib/src/sentry_exception_factory.dart
@@ -22,6 +22,7 @@ class SentryExceptionFactory {
   }) {
     var throwable = exception;
     Mechanism? mechanism;
+
     bool? snapshot;
     if (exception is ThrowableMechanism) {
       throwable = exception.throwable;